metamixr: 'MetaMix': Bayesian Meta-Analytic Mixture Models

Fits Bayesian random-effects meta-analytic mixture models in which the true study effects are drawn from a mixture of normal components with component-specific means and between-study heterogeneity. The mixture model can be combined with step-function selection models (Vevea and Hedges, 1995, doi:10.1007/BF02294384 ) to adjust for publication bias, either with one selection model for all studies or with separate selection models for groups of studies. Models are estimated by Hamiltonian Monte Carlo using 'rstan' and returned as 'stanfit' objects, so posterior summaries, posterior predictive checks and marginal likelihoods for model comparison (via 'bridgesampling') are readily available. Also includes a data simulator and the nudging meta-analysis data of Mertens et al. (2022) doi:10.1073/pnas.2107346118 .

Details

The package implements Bayesian meta-analytic mixture models with publication bias correction. Three model fitting functions all return a stanfit object:

re_mix()

random-effects meta-analytic mixture model without publication bias adjustment (the ordinary random-effects meta-analysis when M = 1).

sel_mix()

the same mixture model combined with a step-function selection model to adjust for publication bias.

sel_flexpb()

the selection model but different selection processes can be estimated for different groups of studies.

sim_mix() simulates data from these models, and mertens_nudge contains the nudging meta-analysis of Mertens et al. (2022) used in the vignettes.

References

Maier, M. (2026). Addressing heterogeneity with Bayesian meta-analytic mixture modelling. PsyArXiv. doi:10.31234/osf.io/nkyqm_v2

Maier, M., Bartoš, F., & Wagenmakers, E.-J. (2023). Robust Bayesian meta-analysis: Addressing publication bias with model-averaging. Psychological Methods, 28(1), 107–122. doi:10.1037/met0000405

Stan Development Team (2024). RStan: the R interface to Stan. https://mc-stan.org/

See also

Author

Maintainer: Maximilian Maier maximilianmaier0401@gmail.com (ORCID)

Other contributors:

Trustees of Columbia University [copyright holder]